+++ tests/conftest.py
@@ -2,6 +2,7 @@
 import logging
 from os.path import abspath, dirname, join
 
+import geopandas as gpd
 import pytest
 import xarray as xr
 from hydromt.cli.cli_utils import parse_config
@@ -77,5 +78,13 @@ def floodplain1d_testdata():
         join(TESTDATADIR, "floodplain_layers.nc"),
         lock=False,
         mode="r",
+    )
+    return data
+
+
+@pytest.fixture()
+def rivers1d():
+    data = gpd.read_file(
+        join(TESTDATADIR, "rivers.geojson"),
     )
     return data